59,166 posts with emotions happiness and sadness
Meu Querido Diário or MQD is a social network where users write texts related to their everyday experiences, like records in a personal diary that can contain thoughts, experiences or their feelings.

This dataset contains 32,244 posts associated with happiness class and 26,922$ with sadness class.
We collected a total of 59,166 posts with emotions happiness and sadness associated, randomly selected from users in the age between 13 and 99, without gender distinction. Caution was taken to assure identifying information stripped, such that it cannot be linked to the person who wrote it.